requirements.txt
slack_sdk
discord.py
slack_bolt
fastapi
uvicorn
requests
